Migrate HTTP types to separate file
